Fix non-literal password string comparison
When using `[[ $x != $y ]]`, the RHS argument is evaluated as a pattern
instead of a string literal, unless it is quoted. This causes the check
to incorrectly fail when using a randomly-generated password containing
a backslash. For example, the password `test\%` never passes the check.

This commit also adds the `-r` argument to a couple password `read`
calls that were missing it so that backslashes are not evaluated as
escape characters.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>